GT2ModelTool: Load material setup from the JSON file when importing from OBJ, falling back to material name parsing if not found to support advanced users who want to use the naming scheme instead, and to support OBJs from v0.1. Shadow materials must still be named correctly currently.

+++ GT2ModelTool/GT2ModelTool/ExportMetadata/MaterialMetadata.cs
@@ -10,7 +10,7 @@ namespace GT2.ModelTool.ExportMetadata
         public bool IsUntextured { get; set; }
 
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-        public int? PaletteIndex { get; set; }
+        public ushort? PaletteIndex { get; set; }
 
         public int RenderOrder { get; set; }
